+++ internal/transport/service.go
@@ -1,10 +1,8 @@
-package game
+package transport
 
-import (
-	"2sp/pkg/protocol"
-)
+import "2sp/pkg/protocol"
 
-type Transport interface {
+type Service interface {
 	SubscribeToMessages(room *protocol.Room) (*MessagesSubscription, error)
 	PublishUnencryptedMessage(room *protocol.Room, payload []byte) error
 	PublishPublicMessage(room *protocol.Room, payload []byte) error
